gpsd/gpsd.c: Fix 32-bit warning.

+++ gpsd/gpsd.c
@@ -616,8 +616,10 @@ static ssize_t throttled_write(struct subscriber_t *sub, char *buf,
 
     if (LOG_CLIENT <= context.errout.debug) {
         if (isprint((unsigned char) buf[0])) {
+            // (long)len for 32-bit compatibility
             GPSD_LOG(LOG_CLIENT, &context.errout,
-                     "=> client(%d) len %ld: %s\n", sub_index(sub), len, buf);
+                     "=> client(%d) len %ld: %s\n",
+                     sub_index(sub), (long)len, buf);
         } else {
             char *cp, buf2[MAX_PACKET_LENGTH * 3];
             buf2[0] = '\0';
@@ -626,7 +628,7 @@ static ssize_t throttled_write(struct subscriber_t *sub, char *buf,
                                "%02x", (unsigned int)(*cp & 0xff));
             GPSD_LOG(LOG_CLIENT, &context.errout,
                      "=> client(%d) len %ld: =%s\n",
-                     sub_index(sub), len, buf2);
+                     sub_index(sub), (long)len, buf2);
         }
     }
